+++ standalone/src/scanner_configuration.cpp
@@ -17,49 +17,52 @@
 
 namespace psen_scan_v2_standalone
 {
+namespace configuration
+{
 bool ScannerConfiguration::isValid() const
 {
   return host_ip_ && scanner_ip_ && scan_range_;
 }
 
-uint32_t ScannerConfiguration::hostIp() const
+uint32_t configuration::ScannerConfiguration::hostIp() const
 {
   return *host_ip_;
 }
 
-uint16_t ScannerConfiguration::hostUDPPortData() const
+uint16_t configuration::ScannerConfiguration::hostUDPPortData() const
 {
   return host_data_port_;
 }
 
-uint16_t ScannerConfiguration::hostUDPPortControl() const
+uint16_t configuration::ScannerConfiguration::hostUDPPortControl() const
 {
   return host_control_port_;
 }
 
-uint32_t ScannerConfiguration::clientIp() const
+uint32_t configuration::ScannerConfiguration::clientIp() const
 {
   return *scanner_ip_;
 }
 
-uint16_t ScannerConfiguration::scannerDataPort() const
+uint16_t configuration::ScannerConfiguration::scannerDataPort() const
 {
   return scanner_data_port_;
 }
 
-uint16_t ScannerConfiguration::scannerControlPort() const
+uint16_t configuration::ScannerConfiguration::scannerControlPort() const
 {
   return scanner_control_port_;
 }
 
-const DefaultScanRange& ScannerConfiguration::scanRange() const
+const DefaultScanRange& configuration::ScannerConfiguration::scanRange() const
 {
   return *scan_range_;
 }
 
-bool ScannerConfiguration::diagnosticsEnabled() const
+bool configuration::ScannerConfiguration::diagnosticsEnabled() const
 {
   return diagnostics_enabled_;
 }
 
+}  // namespace configuration
 }  // namespace psen_scan_v2_standalone